Create workspace runner for cypress with separate mount
