+++ packages/mount/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@cypress-lit/mount",
+  "description": "",
+  "version": "0.0.1",
+  "author": "Simon Reilly",
+  "devDependencies": {
+    "@cypress/mount-utils": ">=2.1.0",
+    "cypress": "^10.10.0",
+    "lit": "^2.3.1",
+    "prettier": "^2.7.1",
+    "rollup": "^3.2.5",
+    "typescript": "^4.6.4",
+    "vite": "^3.1.0"
+  },
+  "files": [
+    "dist/**/*",
+    "types/**/*"
+  ],
+  "keywords": [],
+  "license": "MIT",
+  "main": "./dist/mount.js",
+  "module": "./dist/mount.mjs",
+  "peerDependencies": {
+    "@cypress/mount-utils": ">=2.1.0",
+    "cypress": ">=10.6.0",
+    "svelte": ">=3.0.0"
+  },
+  "scripts": {
+    "build": "vite build",
+    "compile": "tsc --build",
+    "cypress": "cypress",
+    "lint": "prettier './{src,cypress}/**/*.{js,ts}' --write"
+  },
+  "types": "types/index.d.ts"
+}
