+++ packages/mount/src/index.ts
@@ -1,0 +1,1 @@
+export { mount } from "./mount";
