+++ tests/integration/cypress/component/web-component.cy.ts
@@ -1,39 +1,37 @@
 import "../../src";
-import { html } from "lit";
-import { LitCounter } from "../../src/counter-lit";
+import { WebCounter } from "../../src";
 
-describe("Lit mount", () => {
+describe("Web Component mount", () => {
   it("mounts", () => {
-    cy.mount<"counter-lit">(html`<counter-lit></counter-lit>`);
-    cy.get("counter-lit").shadow().contains("h1", "Count is 0");
+    cy.mount<"counter-wc">(`<counter-wc></counter-wc>`);
+    cy.get("counter-wc").shadow().contains("h1", "Count is 0");
+  });
+
+  it("accepts props", () => {
+    cy.mount<"counter-wc">(`<counter-wc count=${42}></counter-wc>`);
+    cy.get("counter-wc").shadow().as("shadow");
+
+    cy.get("@shadow").contains("h1", "Count is 42");
   });
 
   it("reacts to state changes", () => {
-    cy.mount<"counter-lit">(html`<counter-lit></counter-lit>`);
+    cy.mount<"counter-wc">(`<counter-wc></counter-wc>`);
 
-    cy.get("counter-lit").shadow().as("shadow");
-
+    cy.get("counter-wc").shadow().as("shadow");
     cy.get("@shadow").contains("h1", "Count is 0");
     cy.get("@shadow").find("button").click();
     cy.get("@shadow").contains("h1", "Count is 1");
   });
 
-  it("accepts props", () => {
-    cy.mount<"counter-lit">(html`<counter-lit .count=${42}></counter-lit>`);
-    cy.get("counter-lit").shadow().as("shadow");
-
-    cy.get("@shadow").contains("h1", "Count is 42");
-  });
-
   it("can pass emitters as spies", () => {
-    cy.mount<"counter-lit">(
-      html`<counter-lit
-        .count=${42}
-        .clicked=${cy.spy().as("onClickedSpy")}
-      ></counter-lit>`
+    cy.mount<"counter-wc">(
+      `<counter-wc
+        count=${42}
+      ></counter-wc>`,
+      { properties: { clicked: cy.spy().as("onClickedSpy") } }
     );
 
-    cy.get("counter-lit").shadow().as("shadow");
+    cy.get("counter-wc").shadow().as("shadow");
 
     cy.get("@shadow").contains("h1", "Count is 42");
     cy.get("@shadow").find("button").click();
@@ -42,30 +40,36 @@ describe("Lit mount", () => {
 
   describe("slotting", () => {
     it("can slot HTMLElements", () => {
-      cy.mount<"counter-lit">(
-        html`<counter-lit .count=${42} .clicked=${cy.spy().as("onClickedSpy")}>
+      cy.mount<"counter-wc">(
+        `<counter-wc
+          count=${42}
+          clicked=${cy.spy().as("onClickedSpy")}
+        >
           <div class="div-slotted">
             <p>Rendered</p>
           </div>
-        </counter-lit>`
+        </counter-wc>`
       );
 
-      cy.get("counter-lit").shadow().as("shadow");
+      cy.get("counter-wc").shadow().as("shadow");
 
       cy.get("@shadow").get(".div-slotted").find("p").contains("Rendered");
     });
 
     it("can slot other web components", () => {
-      cy.mount<"counter-lit">(
-        html`<counter-lit .count=${42} .clicked=${cy.spy().as("onClickedSpy")}>
-          <counter-lit .count=${99}></counter-lit>
-        </counter-lit>`
+      cy.mount<"counter-wc">(
+        `<counter-wc
+          count=${42}
+          clicked=${cy.spy().as("onClickedSpy")}
+        >
+          <counter-wc count=${99}></counter-wc>
+        </counter-wc>`
       );
 
-      cy.get("counter-lit").shadow().as("shadow");
+      cy.get("counter-wc").shadow().as("shadow");
       cy.get("@shadow").contains("h1", "Count is 42");
       cy.get("@shadow")
-        .get("counter-lit")
+        .get("counter-wc")
         .shadow()
         .contains("h1", "Count is 99");
     });
@@ -73,17 +77,15 @@ describe("Lit mount", () => {
 
   describe("wrapping", () => {
     it("component is instance of web component", () => {
-      cy.mount<"counter-lit">(html`<counter-lit></counter-lit>`).then(
-        ({ component }) => {
-          expect(component).to.be.instanceOf(LitCounter);
-        }
-      );
+      cy.mount<"counter-wc">(`<counter-wc></counter-wc>`).then((component) => {
+        expect(component[0]).to.be.instanceOf(WebCounter);
+      });
     });
   });
 
   context("log", () => {
     it("displays component name in mount log", () => {
-      cy.mount<"counter-lit">(html`<counter-lit .count=${42}></counter-lit>`);
+      cy.mount<"counter-wc">(`<counter-wc count=${42}></counter-wc>`);
 
       cy.wrap(Cypress.$(window.top.document.body)).within(() =>
         cy
@@ -93,13 +95,13 @@ describe("Lit mount", () => {
           .within(() =>
             cy
               .get(".command-name-mount")
-              .should("contain", "mount<counter-lit ... />")
+              .should("contain", "mount<counter-wc ... />")
           )
       );
     });
 
     it("does not display mount log", () => {
-      cy.mount<"counter-lit">(html`<counter-lit .count=${42}></counter-lit>`, {
+      cy.mount<"counter-wc">(`<counter-wc .count=${42}></counter-wc>`, {
         log: false,
       });
