+++ tests/integration/cypress/support/component.ts
@@ -19,7 +19,7 @@ import "./commands";
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
 
-import { mount } from "./lit";
+import { mount } from "@cypress-lit/mount";
 
 // Augment the Cypress namespace to include type definitions for
 // your custom command.
