+++ tests/integration/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@cypress-lit/integration",
+  "version": "0.0.0",
+  "dependencies": {
+    "lit": "^2.3.1"
+  },
+  "devDependencies": {
+    "@cypress-lit/mount": "workspace:^0.0.1",
+    "cypress": "^10.10.0",
+    "prettier": "^2.7.1",
+    "typescript": "^4.6.4",
+    "vite": "^3.1.0"
+  },
+  "exports": {
+    ".": "./dist/counter-lit.es.js"
+  },
+  "files": [
+    "dist",
+    "types"
+  ],
+  "main": "dist/counter-lit.es.js",
+  "private": true,
+  "scripts": {
+    "cypress": "cypress",
+    "dev": "vite",
+    "lint": "prettier './{src,cypress}/**/*.{js,ts}' --write"
+  },
+  "type": "module",
+  "types": "types/counter-lit.d.ts"
+}
