+++ tests/integration/src/counter-lit.ts
