+++ tests/integration/src/counter-wc.ts
