+++ tests/integration/src/vite-env.d.ts
