--- vite.config.ts
@@ -1,14 +1,0 @@
-import { defineConfig } from 'vite'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  build: {
-    lib: {
-      entry: 'src/index.ts',
-      formats: ['es']
-    },
-    rollupOptions: {
-      external: /^lit/
-    }
-  }
-})
